@@ -11,11 +11,14 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.2"
     },
     "autoload": {
         "psr-0": {
             "RestService": ""
         }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^7.5"
     }
 }
