@@ -4,7 +4,7 @@ namespace Test\Synthetic;
 
 use RestService\Server;
 
-class RouteTest extends \PHPUnit_Framework_TestCase
+class RouteTest extends \PHPUnit\Framework\TestCase
 {
 
     public function testAllRoutesClosures()
